feat(tearchan): Update gfx-hal to v0.6.0.

+++ src/core/graphic/hal/index_buffer.rs
@@ -134,7 +134,7 @@ fn create_index_buffer<B: Backend>(
         .iter()
         .enumerate()
         .position(|(idx, mem_type)| {
-            buffer_req.type_mask & (1 << idx) as u64 != 0
+            buffer_req.type_mask & (1 << idx) as u32 != 0
                 && mem_type
                     .properties
                     .contains(memory::Properties::CPU_VISIBLE)
